pyproject.toml
[tool.poetry]
name = "discover-service-k8s"
version = "0.0.1"
description = "An implementation of the Discover Web Service using Kubernetes"
authors = ["Charles Landau <clandau@guidehouse.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
kubernetes = "^23.3.0"
Jinja2 = "^3.1.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"
py-healthcheck = "^1.10.1"
fastapi-login = "^1.8.0"
typer = "^0.4.1"
boto3 = "^1.22.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
